requirements.txt
azure-storage-blob>=12.19.0
azure-search-documents>=11.4.0
openai>=1.0.0
langchain>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
